cleanup
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@30950 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/agp_validate/ContextValidator.cpp
@@ -155,7 +155,7 @@ void CAgpContextValidator::x_OnGapLine(
 {
   int i = gap.type;
   if(gap.linkage==CGapVal::LINKAGE_yes) i+= CGapVal::GAP_count;
-  NCBI_ASSERT( i < sizeof(m_GapTypeCnt)/sizeof(m_GapTypeCnt[0]),
+  NCBI_ASSERT( i < (int)sizeof(m_GapTypeCnt)/sizeof(m_GapTypeCnt[0]),
     "m_GapTypeCnt[] index out of bounds" );
   m_GapTypeCnt[i]++;
   m_GapCount++;
